pyproject.toml
[tool.poetry]
name = "mnist-mlops"
version = "0.0.1"
description = "mlops pipeline with tensor-flow for the mnist dataset using concepts of mlops and digital image processing"
authors = ["Diego Rodrigues Medeiros[diego25rn@gmail.com]"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
opencv-python = "^4.6.0"
tensorflow = "^2.9.1"
seaborn = "^0.11.2"
keras-tuner = "^1.1.2"
tensorflow-datasets = "^4.6.0"
plotly = "^5.9.0"
streamlit = "^1.10.0"
idx2numpy = "^1.2.3"
wandb = "^0.12.21"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
